% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{background_risk}
\alias{background_risk}
\title{Compute background risk for T21 as a function of mother's age and gestational age.}
\usage{
background_risk(age_mother, gestage)
}
\arguments{
\item{age_mother}{Age of the mother (in years)}

\item{gestational}{age (in days) at the time measurements were taken}
}
\value{
Computes prior risk for Trisomy 21 based on the table in Snijders et al. (1999)
}
\description{
Compute background risk for T21 as a function of mother's age and gestational age.
}
