% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{crl2gestage}
\alias{crl2gestage}
\title{Reverse function of gestage2crl}
\usage{
crl2gestage(crl)
}
\arguments{
\item{gestage}{Gestational age (in days)}
}
\value{
Crown-rump-length (in mm)
}
\description{
Reverse function of gestage2crl
}
\seealso{
[trisomy21risk::gestage2crl]
}
