% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{doe}
\alias{doe}
\title{Compute the DoE (Degrees of Extremeness) of an observation}
\usage{
doe(y, y_pred, ref_lower, ref_upper)
}
\arguments{
\item{y}{The observed value}

\item{y_pred}{This is the model based expected value}

\item{ref_lower}{Lower value of the reference band, i.e. the 5% quantile value}

\item{ref_upper}{Upper value of the reference band, i.e. the 95% quantile value}
}
\description{
Compute the DoE (Degrees of Extremeness) of an observation
}
\references{
Merz, E., C. Thode, B. Eiben, and S. Wellek. 2016. “Prenatal Risk Calculation (PRC) 3.0: An Extended Doe-Based First-Trimester Screening Algorithm Allowing for Early Blood Sampling.” Ultrasound International Open 2: E19–E26. doi:10.1055/s-0035-1569403.
}
