% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{gestage2crl}
\alias{gestage2crl}
\title{Translate from age to CRL using the Wikipedia formula https://en.wikipedia.org/wiki/Crown-rump_length,
where CRL is measured in mm and gestational age in days.}
\usage{
gestage2crl(gestage)
}
\arguments{
\item{gestage}{Gestational age (in days)}
}
\value{
Crown-rump-length (in mm)
}
\description{
Gestational age = (CRL x 1.04)^0.5 x 8.05 + 23.7
}
\details{
Values on FMF page https://fetalmedicine.org/research/pregnancyDating with JavaScript https://fetalmedicine.org/assets/7a3311bf/js/base.min.js?v=51860
The page attributes https://www.ncbi.nlm.nih.gov/pubmed/1182090

gaFromCrl:function(t,e,n){
  return e=e||45,n=n||84,t=this.parseFloat(t),null===t||t<e||t>n?null:23.53+8.052*Math.sqrt(1.037*t)
}

We take the formula from http://journals.sagepub.com/doi/pdf/10.1179/174313409X448543
Gestational age = (CRL x 1.037)^0.5 x 8.052 + 23.73
}
\references{
http://journals.sagepub.com/doi/pdf/10.1179/174313409X448543
}
