% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{logbetahCG_corr}
\alias{logbetahCG_corr}
\title{Compute weight corrected logbetahCG as done in Merz et al. (2016)}
\usage{
logbetahCG_corr(betahCG, weight)
}
\arguments{
\item{betahCG}{Concentration of beta-hCG in maternal serum (in miU/ml)}

\item{weight}{Weight of the mother (in kg)}
}
\value{
log(beta-hCG) concentration corrected for the weight of the mother
}
\description{
Compute weight corrected logbetahCG as done in Merz et al. (2016)
}
