% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{logPAPPA}
\alias{logPAPPA}
\title{Manual least squares fit to the Merz et al. (2016) data using a 10th order polynomial}
\usage{
logPAPPA(crlbe)
}
\arguments{
\item{crlbe}{crown rump length (in mm) at the time of the serum sampling}
}
\value{
The expected value of logPAPPA (from a 10th order polynomial fitted to the data of Merz et al.)
}
\description{
Manual least squares fit to the Merz et al. (2016) data using a 10th order polynomial
}
\references{
Merz, E., C. Thode, B. Eiben, and S. Wellek. 2016. “Prenatal Risk Calculation (PRC) 3.0: An Extended Doe-Based First-Trimester Screening Algorithm Allowing for Early Blood Sampling.” Ultrasound International Open 2: E19–E26. doi:10.1055/s-0035-1569403.
}
