% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{logPAPPA_corr}
\alias{logPAPPA_corr}
\title{Compute weight corrected logPAPPA as done in Merz et al. (2016)}
\usage{
logPAPPA_corr(pappa, weight)
}
\arguments{
\item{pappa}{Concentration of PAPP-A in maternal serum (in miU/ml)}

\item{weight}{Weight of the mother (in kg)}
}
\value{
log(PAPP-A) concentration corrected for the weight of the mother
}
\description{
Compute weight corrected logPAPPA as done in Merz et al. (2016)
}
