% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{ppost}
\alias{ppost}
\title{Compute posterior distribution for an event using the formula:}
\usage{
ppost(lr_joint, pi = 1/100)
}
\arguments{
\item{lr_joint}{likelihood ratio}

\item{pi}{prior probability}
}
\value{
The posterior odds as \eqn{1/(1+1/lr_joint*(1-pi)/pi)}
}
\description{
posterior odds = likelihood ratio * prior odds
}
