% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{trisomy21}
\alias{trisomy21}
\title{Function to compute the trisomy21 risk based on 3 biomarkers (NT, PAPP-A, betahCG)
and visibility of the nasal bone on the ultrasound. Several covariates (age of mother, weight of mother, CRL at measurement)
are taken into account.}
\usage{
trisomy21(age, weight, crlbe, crl, nt, pappa, betahCG, nasalbone, background)
}
\arguments{
\item{age}{Age of the mother (in years)}

\item{weight}{Weight of the mother (in kg)}

\item{crlbe}{Crown-rump-length (in mm) at the time of the PAPP-A and betahCG measurements}

\item{crl}{Crown-rump-length (in mm) at the time of the NT measurement}

\item{pappa}{Measured value of PAPP-A in maternal serum (in miU/ml)}

\item{betahCG}{Measured value of betahCG in maternal serum (in miU/ml)}

\item{nasalbone}{Visibility of a nasal bone on the ultrasound (Boolean)}

\item{background}{Background probability for T21}
}
\value{
Posterior risk of T21.
}
\description{
Function to compute the trisomy21 risk based on 3 biomarkers (NT, PAPP-A, betahCG)
and visibility of the nasal bone on the ultrasound. Several covariates (age of mother, weight of mother, CRL at measurement)
are taken into account.
}
\references{
Merz, E., C. Thode, B. Eiben, and S. Wellek. 2016. “Prenatal Risk Calculation (PRC) 3.0: An Extended Doe-Based First-Trimester Screening Algorithm Allowing for Early Blood Sampling.” Ultrasound International Open 2: E19–E26. doi:10.1055/s-0035-1569403.
}
