\name{trisomy21risk-package}
\alias{trisomy21risk-package}
\alias{trisomy21risk}
\docType{package}
\title{
\packageTitle{trisomy21risk}
}
\description{
\packageDescription{trisomy21risk}
}
\details{
  \packageDESCRIPTION{trisomy21risk}
}
\author{
\packageAuthor{trisomy21risk}
}
\references{
  Merz, E., C. Thode, B. Eiben, and S. Wellek. 2016. “Prenatal Risk
  Calculation (PRC) 3.0: An Extended Doe-Based First-Trimester Screening
  Algorithm Allowing for Early Blood Sampling.” Ultrasound International
  Open 2: E19–E26. doi:10.1055/s-0035-1569403.
}

