% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t21.R
\name{doe_inv}
\alias{doe_inv}
\title{Inverse function of the DoE (Degrees of Extremeness), i.e. compute observed value given DoE}
\usage{
doe_inv(doe, y_pred, ref_lower, ref_upper)
}
\arguments{
\item{doe}{The observed degree of extremeness value}

\item{y_pred}{This is the model based expected value}

\item{ref_lower}{Lower value of the reference band, i.e. the 5th percentile value}

\item{ref_upper}{Upper value of the reference band, i.e. the 95th percentile value}
}
\value{
The corresponding y value
}
\description{
Inverse function of the DoE (Degrees of Extremeness), i.e. compute observed value given DoE
}
\seealso{
[trisomy21risk::doe]
}
