\name{trisomy21risk-package}
\alias{trisomy21risk-package}
\alias{trisomy21risk}
\docType{package}
\title{
\packageTitle{trisomy21risk}
}
\description{
\packageDescription{trisomy21risk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trisomy21risk}
\packageIndices{trisomy21risk}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trisomy21risk}

Maintainer: \packageMaintainer{trisomy21risk}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
